+++ packages/lightning-mock/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/lightning-mock
 
+## 2.0.2
+
+### Patch Changes
+
+- @openfn/engine-multi@1.1.2
+
 ## 2.0.1
 
 ### Patch Changes